+++ src/Backdrop/Backdrop.stories.tsx
@@ -13,7 +13,7 @@
  * limitations under the License.                                           *
  * ======================================================================== */
 import React from 'react';
-import { StoryFn, Meta } from '@storybook/react';
+import { StoryFn, Meta } from '@storybook/react-webpack5';
 import { Theme } from '@mui/material';
 
 import Backdrop from './Backdrop';