+++ src/Breadcrumbs/Breadcrumbs.stories.tsx
@@ -14,7 +14,7 @@
  * ======================================================================== */
 
 import React from 'react';
-import { StoryFn, Meta } from '@storybook/react';
+import { StoryFn, Meta } from '@storybook/react-webpack5';
 import IconInformation from '@hcl-software/enchanted-icons/dist/carbon/es/information';
 
 import {
@@ -38,9 +38,10 @@ export default {
   argTypes: {
     disabled: {
       description: 'If `true`, the component is disabled.',
+      control: 'boolean',
       table: {
         defaultValue: {
-          summary: false,
+          summary: 'false',
         },
       },
     },