+++ src/TextField/TextField.stories.tsx
@@ -14,12 +14,12 @@
  * ======================================================================== */
 
 import React from 'react';
-import { StoryFn, Meta } from '@storybook/react';
+import { StoryFn, Meta } from '@storybook/react-webpack5';
 import CaretDownIcon from '@hcl-software/enchanted-icons/dist/carbon/es/caret--down';
 import InformationIcon from '@hcl-software/enchanted-icons/dist/carbon/es/information';
 
 import InputAdornment from '@mui/material/InputAdornment';
-import { Box } from '@mui/material';
+import { Box, SvgIconProps } from '@mui/material';
 import TextField from './TextField';
 import Button from '../Button';
 
@@ -76,7 +76,7 @@ export default {
     enableHelpHoverEffect: {
       control: 'boolean',
       table: {
-        defaultValue: { summary: false },
+        defaultValue: { summary: 'false' },
       },
       description: 'If true, the helper icon displays a gray background when hovered.',
     },
@@ -100,45 +100,54 @@ export default {
     },
     error: {
       description: 'If `true`, the error state will be enabled.',
+      control: 'boolean',
       table: {
         disable: TextField.defaultProps?.error,
       },
     },
     hiddenLabel: {
       description: 'If `true`, the label will hide.',
+      control: 'boolean',
       table: {
         defaultValue: { summary: TextField.defaultProps?.hiddenLabel },
       },
     },
     nonEdit: {
       description: 'If `true`, the component is only ready. No interactions are possible..',
+      control: 'boolean',
       table: {
         defaultValue: { summary: TextField.defaultProps?.nonEdit },
       },
     },
     disabled: {
       description: 'If `true`, the component is disabled.',
+      control: 'boolean',
       table: {
         defaultValue: { summary: TextField.defaultProps?.disabled },
       },
     },
     required: {
       description: 'If `true`, the `input` element is required.',
+      control: 'boolean',
       table: {
         defaultValue: { summary: TextField.defaultProps?.required },
       },
     },
     sx: {
       description: 'The system prop that allows defining system overrides as well as additional CSS styles.',
+      control: false,
     },
     fullWidth: {
       description: 'If true, the input will take up the full width of its container.',
+      control: 'boolean',
     },
     autoFocus: {
       description: 'If true, the input element is focused during the first mount.',
+      control: 'boolean',
     },
     multiline: {
       description: 'If true, a textarea element is rendered instead of an input.',
+      control: 'boolean',
     },
     endAdornmentAction: {
       control: false,
@@ -172,7 +181,7 @@ export default {
 const Template: StoryFn<typeof TextField> = (args) => {
   const [value, setValue] = React.useState(args.value ? args.value : '');
 
-  let customIcon: React.ComponentType<React.SVGProps<SVGSVGElement>> | undefined;
+  let customIcon: React.ComponentType<SvgIconProps> | undefined;
   switch (args.customIcon as unknown as string) {
     case 'CaretDownIcon':
       customIcon = CaretDownIcon;