+++ src/ToggleButtonGroup/ToggleButtonGroup.stories.tsx
@@ -14,7 +14,7 @@
  * ======================================================================== */
 
 import React from 'react';
-import { StoryFn, Meta } from '@storybook/react';
+import { StoryFn, Meta } from '@storybook/react-webpack5';
 import IconAdd from '@hcl-software/enchanted-icons/dist/carbon/es/add';
 import { Grid } from '@mui/material';
 
@@ -40,10 +40,12 @@ export default {
     },
     disabled: {
       if: { arg: 'interactive' },
+      control: 'boolean',
       description: 'Disables the ToggleButton',
     },
     exclusive: {
       if: { arg: 'interactive' },
+      control: 'boolean',
       description: 'If value is exclusive or not',
     },
     orientation: {