+++ src/composite_components/MultipleSelectChip/MultipleSelectChip.stories.tsx
@@ -14,13 +14,13 @@
  * ======================================================================== */
 
 import React from 'react';
-import { StoryFn, Meta } from '@storybook/react';
+import { StoryFn, Meta } from '@storybook/react-webpack5';
 import CloseIcon from '@hcl-software/enchanted-icons/dist/carbon/es/close';
 import CheckmarkIcon from '@hcl-software/enchanted-icons/dist/carbon/es/checkmark';
 import CaretDownIcon from '@hcl-software/enchanted-icons/dist/carbon/es/caret--down';
 import InformationIcon from '@hcl-software/enchanted-icons/dist/carbon/es/information';
 import WarningIcon from '@hcl-software/enchanted-icons/dist/carbon/es/warning--alt';
-import { Alert, Box } from '@mui/material';
+import { Alert, Box, SvgIconProps } from '@mui/material';
 import Chip from '../../Chip/Chip';
 import MultipleSelectChip from './MultipleSelectChip';
 import MenuItem from '../../Menu/MenuItem';
@@ -39,25 +39,35 @@ export default {
   title: 'Inputs/MultipleSelectChip',
   component: MultipleSelectChip,
   argTypes: {
+    size: {
+      control: { type: 'radio' },
+      options: ['small', 'medium'],
+      description: 'The size of the component.',
+      table: { type: { summary: 'string' } },
+    },
     error: {
       control: 'boolean',
       table: {
         defaultValue: {
-          summary: false,
+          summary: 'false',
         },
+        type: { summary: 'boolean' },
       },
       description: 'Indicates the combobox value is invalid.',
     },
     emptyOptions: {
       control: 'boolean',
+      table: { type: { summary: 'boolean' } },
       description:
         'Adds freeSolo and empty options array to the MultiSelectChip which would allow it to accept free undetermined values (e.g. any word or phrase)',
     },
     helperIconTooltip: {
+      control: 'text',
       table: {
         defaultValue: {
           summary: 'Some information about that component.',
         },
+        type: { summary: 'string' },
       },
       description:
         'Tooltip text hovering on ? mark for MultiSelectChip component',
@@ -84,60 +94,71 @@ export default {
         defaultValue: {
           summary: TooltipPlacement.BOTTOM,
         },
+        type: { summary: 'string' },
       },
     },
     required: {
       control: 'boolean',
       table: {
         defaultValue: {
-          summary: true,
+          summary: 'true',
         },
+        type: { summary: 'boolean' },
       },
       description:
         'Indicates that multiSelectchip is required field if it is true',
     },
     label: {
+      control: 'text',
       table: {
         defaultValue: {
           summary: 'Label',
         },
+        type: { summary: 'string' },
       },
       description: 'Label for MultiSelectChip component',
     },
     helperText: {
+      control: 'text',
       table: {
         defaultValue: {
           summary: 'Helper text',
         },
+        type: { summary: 'string' },
       },
       description: 'The helper text content.',
     },
     sx: {
       description:
         'The system prop that allows defining system overrides as well as additional CSS styles.',
+      control: false,
     },
     enableHelpHoverEffect: {
       control: 'boolean',
       table: {
-        defaultValue: { summary: false },
+        defaultValue: { summary: 'false' },
+        type: { summary: 'boolean' },
       },
       description: 'If true, the helper icon displays a gray background when hovered.',
     },
     placeholder: {
+      control: 'text',
       description:
         'The short hint displayed in the input before the user enters a value.',
       table: {
         defaultValue: {
           summary: 'Placeholder',
         },
+        type: { summary: 'string' },
       },
     },
     focused: {
       control: 'boolean',
       table: {
         defaultValue: {
-          summary: false,
+          summary: 'false',
         },
+        type: { summary: 'boolean' },
       },
       description: 'If `true`, the component is displayed in focused state.',
     },
@@ -145,8 +166,9 @@ export default {
       control: 'boolean',
       table: {
         defaultValue: {
-          summary: false,
+          summary: 'false',
         },
+        type: { summary: 'boolean' },
       },
       description: 'If `true`, the label is hidden.',
     },
@@ -154,43 +176,54 @@ export default {
       control: 'boolean',
       table: {
         defaultValue: {
-          summary: false,
+          summary: 'false',
         },
+        type: { summary: 'boolean' },
       },
       description: 'If `true` value cannot be editable',
     },
     actionProps: {
       description: 'Contains the action props and key-value',
+      control: false,
     },
     disabled: {
       description: 'If true, the component is disabled.',
+      control: 'boolean',
+      table: { type: { summary: 'boolean' } },
     },
     fullWidth: {
+      control: 'boolean',
       description:
         'If true, the input will take up the full width of its container.',
       table: {
         defaultValue: {
-          summary: false,
+          summary: 'false',
         },
+        type: { summary: 'boolean' },
       },
     },
     filterSelectedOptions: {
+      control: 'boolean',
       description: 'If true, hide the selected options from the list box.',
       table: {
         defaultValue: {
-          summary: true,
+          summary: 'true',
         },
+        type: { summary: 'boolean' },
       },
     },
     options: {
       description: 'Array of options.',
+      control: 'object',
     },
     id: {
+      control: 'text',
       description: 'Id of the component.',
       table: {
         defaultValue: {
           summary: 'id',
         },
+        type: { summary: 'string' },
       },
     },
     clearIcon: {
@@ -217,7 +250,7 @@ export default {
       control: false,
       table: {
         defaultValue: {
-          summary: false,
+          summary: 'false',
         },
       },
       description: 'https://mui.com/material-ui/api/chip/',
@@ -242,7 +275,7 @@ export default {
 const Template: StoryFn<typeof MultipleSelectChip> = (args) => {
   const [value, setValue] = React.useState([top100Films[13], top100Films[10], top100Films[2]]);
 
-  let customIcon: React.ComponentType<React.SVGProps<SVGSVGElement>> | undefined;
+  let customIcon: React.ComponentType<SvgIconProps> | undefined;
   switch (args.customIcon as unknown as string) {
     case 'CaretDownIcon':
       customIcon = CaretDownIcon;